move GenericFloatingPointPredicateUtils from ADT to IR to fix linkage
